Enforce validity period in COP for ECerts/TCerts
This change set verifies that a certificate has neither
expired nor been revoked.  The main change is a relatively
small change in auth.go which calls cfssl code to verify the
cert.

The test case uses a special signing profile which expires after
1 second.  To support this, I had to make a change so that the
enroll and reenroll commands support using a non-default signing profile.
Therefore, I changed what goes across the network of an enroll
and reenroll request.  Prior to this change, it only sends
a CSR.  After this change, it sends JSON with a CSR and an
optional signing profile name (along with other optional fields).

I have notified the SDK folks of this change.

https://jira.hyperledger.org/browse/FAB-145

Change-Id: Idceeedadf9e5a42d995feee5bdba1353b685e9f1
Signed-off-by: Keith Smith <bksmith@us.ibm.com>