[FAB-1470] Fix docker-clean Makefile target
The docker-clean target was not deleting the
Docker images as expected because of the way
variable substition was working.

This patch modifies the Makefile by adding
additional environment variables in order
to fix this issue.

While there were several ways to solve this,
this approach allows for chainging the name
of the project in the future without having
to make multiple changes in the Makefile

Change-Id: I683f1ba5fbc6ac4c9bce9bcb82922c371f2362aa
Signed-off-by: mastersingh <gari.r.singh@gmail.com>