COP UserRegistry Consolidation
Remove UpdateField/GetField and replace with a single
GetUser and UpdateUser call. This will remove the need
to call both UpdateUser and UpdateField when updating a
field that was not initially part of the UserInfo struct
that was defined in the SPI. The UserInfo struct has been
expanded to support more fields.

https://jira.hyperledger.org/browse/FAB-1442

Change-Id: I1181a54a99bea42f4e6688f55c968798acec8025
Signed-off-by: Saad Karim <skarim@us.ibm.com>