COP API simplification
See https://jira.hyperledger.org/browse/FAB-1462 for
a description of what was done in this change set.
This draws clean lines around the client library API
defined in api/client.go verses the REST APIs of what
goes over the network in api/net.go, which defines
the body of POST and responses.  This should help in
generating swagger doc for the REST APIs.

Note that the tcert library at lib/tcert has it's own
API in lib/tcert/api.go also.

Change-Id: Idb4500b3a5dbc04dfeb65d5b79f8b803411d975f
Signed-off-by: Keith Smith <bksmith@us.ibm.com>