FAB-1338: Include all config and cert files
This change ensures that all of the necessary
sample files are included in the cop image that
is needed in order for a client/peer to connect
successfully.

It should be noted that each peer must have it's
own certificate included in the cop_client.json
file and in the correct volume.

Change-Id: I9204f026f9f5cafca1b6994591f2e214c6345d81
Signed-off-by: Latitia M Haskins <latitia.haskins@gmail.com>