Change the default port to 7054
See https://jira.hyperledger.org/browse/FAB-1782
This changes the default listening fabric-ca port from 8888
to 7054 which is the same port as the old membership services.
This was unfortunately not as trivial as you would think because
the default port is buried in cfssl's cli/config.go file with no
way to change the flag set.  So I process the command line args
prior to calling into cfssl's code to see if a "-port" option
was specified.  If it wasn't, then I add "-port 7054" to the
command line, so to cfssl it looks like we always pass in a "-port"
option.

The GetCommandLineOptValue func in util/args.go may also be used
to remove the "-config" option for the cli client in a future change
set.

Change-Id: I6caca6ab9c36a0a11d2069796cbfc18e39969faa
Signed-off-by: Keith Smith <bksmith@us.ibm.com>