TLS testcases and process file names client config
If a file name in the client config file is not absolute,
it must be processed relative to the location of
the config file.

Added testcases to test TLS with the new CLI and config file.

3 scenarios tested:
1) Read from config file directly
2) Read from environment variable
3) Read from CLI flags

Added new tags to EnrollmentRequest struct for reflection. Added
support to be able to provide root CA certificate via command line
option for TLS configuration on client side.

https://jira.hyperledger.org/browse/FAB-1549

Change-Id: I4ffb52ab4b3b32befbc191a42d2a06dbdde5570e
Signed-off-by: Saad Karim <skarim@us.ibm.com>