Intermediate CA server support
https://jira.hyperledger.org/browse/FAB-1465

This is implements the intermediate CA server support for
fabric-ca-server.  If the "-u <parentURL>" option is used for
"fabric-ca-server init" or "fabric-ca-server start" commands and
the CA cert and key have not yet been generated, the server acts
as an intermediate CA and gets it's certificate from the parent server.
This must be done with an identity which has the "hf.IntermediateCA"
attribute.

The main logic to get the server's CA cert and key are in the
getCACertAndKey function in lib/server.go.  It is this function which
checks for the parentServerURL.  Note that in this case, the server
uses its client object to enroll with the parent server, just as a
standalone client does.

Note the addition of the csrAuthCheck function in lib/serverenroll.go file.
If the request is for a CA certificate, the "hf.IntermediateCA" attribute
is required by the caller.

See TestIntermediateServer in lib/server_test.go for the test case.

Change-Id: Ifefc304627b58ec1fdc5fac9ebf5e9415483de4c
Signed-off-by: Keith Smith <bksmith@us.ibm.com>