Docker image with client and server commands
https://jira.hyperledger.org/browse/FAB-2363

This change set builds the hyperledger/fabric-ca image with both the
fabric-ca-server and fabric-ca-client commands.  The default start CMD
starts the server.
This also adds fabric-ca/docker/server/docker-compose.yml file which starts
the server, and moves the previous compose file to
fabric-ca/docker/examples/client-server-flow/docker-compose.yml since it is
really an example.  Allen will push a change set to update the
client-server-flow/docker-compose.yml file as well as updating the
fabric-ca-fvt image to use the new commands.

Change-Id: I6fe27717ed8428ca4963c9241a5ab128b823a3cc
Signed-off-by: Keith Smith <bksmith@us.ibm.com>