Don't print usage message for non-usage errors
If an error occurs that was not related to usage of the
command, then the usage message should not be displayed
rather just the error that occured is shown.

However, if the problem was the usage of the command then
usage message should be printed.

Also, resolves the issue where an error message was being
sent to client twice during revocation.

https://jira.hyperledger.org/browse/FAB-2835
https://jira.hyperledger.org/browse/FAB-2839

Change-Id: I75696ce396eb4d7be09afcff631aa4195527936d
Signed-off-by: Saad Karim <skarim@us.ibm.com>