Handle string slices in config appropriately
See [FAB-2993]

This change set allows a string slice to be specified via
command line flags or environment variables.  For example,
[FAB-2896] needs to specify
a list of file names with the following option:
   --cafiles <CA-config-files>

This also gets rid of some kludges that were previously done
because string slices weren't handled.

Note that there is a bug in viper which I had to work around
in how it handles string slices.  See comments in code.

Note also that I was getting a transient failure with
this change set in the lib package because tests in different
files in the lib package were running in parallel.  So I'm doing
two things to get this to pass CI:
1) Made all servers in lib to start on different ports there was
a timing issue in the test there.
2) Added more test cases and removed some dead code (NewClient)
to increase test coverage.

Change-Id: I23aeb9d39ece403b85a03304f895fc8b31998805
Signed-off-by: Keith Smith <bksmith@us.ibm.com>