[FAB-3050] Document serial number of CSR
See [FAB-3050]

There are two serial numbers which can be confusing.
The required serial number of the certificate is set by the fabric-ca-server.
The optional serial number which is part of the CSR (Certificate Signing Request)
is added to the DN (Distinquished Name) of the issued certificate.
The use case for the latter is that a company CA issues certificates with
the employee serial number as part of the DN.

This change set adds documentation to the client config file describing
the difference.  It also removes support for the CSR serial number from the
server's config since there is no known use case for this.

Change-Id: I73afe850cba48321a0a7a5ec3898f9d9ebd75cfd
Signed-off-by: Keith Smith <bksmith@us.ibm.com>