[FAB-3924] Improve test coverage of lib
This change set improves the coverage of the
lib package to over 85%. Most of the changes
are to test files.

The changes to none test files are either improvments
to debug/error messages or the removal of code that
was no longer being used. All of which help to
improve test coverage.

Change-Id: I91816df990ac446182118c67f0de18440805ad5c
Signed-off-by: Keith Smith <bksmith@us.ibm.com>
Signed-off-by: rennman <eabailey@us.ibm.com>
Signed-off-by: Saad Karim <skarim@us.ibm.com>