[ FAB-4016 ] fix run_fvt_test redirect error
Some of the fvt tests are recording false positives due
to a synchronization issue writing to the log file.
Currently the rc is out of scope of the main script due
to the fact that when the shell opens a pipe to 'tee' it
operates in a new subshell environment, replete with it's
own I/O stream and variable set. The run_fvt_tests script
will be re-written using command substitution to simplify
the return code logic so that the return code is globally visable

Change-Id: Ife8ecca717d00f2e22ca5fbce7173879bb232455
Signed-off-by: Allen Bailey <eabailey@us.ibm.com>