[FAB-4017] Duplicate flags registered for 'Hosts'
Currently there are two places you can specify 'hosts' when
performing an enrollment. One place is to specify it as part
of the CSR and the second place is in the Enrollment
Request API. For example, you see duplicate hosts flags on
client side:

--csr.hosts stringSlice A list of space-separated host names in a
certificate signing request

--enrollment.hosts string Comma-separated host list

It is unclear to a user as to which one to use or which one will
take precedence. Currently the hosts that is part of the Enrollment
API will only take affect and the 'hosts' from the CSR will get
ignored. This change removes 'hosts' from Enrollment
API, and uses the 'hosts' defined in the CSR.

See [FAB-4017] for more info.

Adds test-cases to get util package test coverage to over 85%

Change-Id: I1d910b29cd9aea98db66511da39345e71df9a525
Signed-off-by: Saad Karim <skarim@us.ibm.com>