[FAB-3734] Fix default expiration times
The default expirations were all 8000 hours which is less than 1 year
for all certificates generated by fabric-ca.
This change set adds documentation to the default config file which
is generated from the code pertaining to how to configure certificate
expiration times for 3 types of certificates.  It also changes the
default values for these types.
The 3 types with their default values are as follows:
1) CA root certificate: default of 15 years
2) CA intermediate certificate:  default of 5 years
3) Issued enrollment certificates: default of 1 year
Although there are no standard values, these values are in line with
what others are doing.

This change set also adds appropriate test cases for these 3
types of certificates.

Change-Id: I4b88517b4550233bfbd5a40ae5b203c584e51a17
Signed-off-by: Keith Smith <bksmith@us.ibm.com>