[FAB-4141] Default CA lookup fails
By default, the default CA in a server has no CA
name (empty string). However, the default CA can be
started up with a CA name specified, but if this done
then client request does not get routed to it appropriately.
If the client does not specify a CA name (empty string), the
server will not correctly direct traffic to the default CA.

Code was added to correctly set the default CA name in client
request when a client makes a request to the default CA.

See [FAB-4141] for more info

Change-Id: If12ab1b4861961a3106bb76f1bee64e2b94cedcd
Signed-off-by: Saad Karim <skarim@us.ibm.com>