[FAB-2919] Workaround for panic due to lib bug
This is a workaround for a panic which is due to a libc bug
on some platforms when linking statically.  There is no fix
for this bug, other than using a different library.  This bug
has only been observed so far when using TLS to Postgres, but
there may be other occurrences.  As a workaround, this change
set documents this bug and also provides instructions on how
to link dynamically when building docker images.

Change-Id: I286bb78ba12572b20f09e0380c6926fa346bf251
Signed-off-by: Keith Smith <bksmith@us.ibm.com>