[FAB-5188] Fix password conversion bug
Passwords are stored in the DB as bytes and declared in the
UserRecord structure as a string.  The UserRecord structure is
passed to the sqlx library to store the user's record in the DB.
The sqlx library interprets the "e" character as an exponent and
converts the string, which means it stores an incorrect password in the DB.

The fix is to change the type of the Pass field from "string" to "[]byte"
to prevent the sqlx library from performing this conversion, and
manually casting between string and []byte as needed.

The TestSpecialPassword test case was added to test this scenario.

Change-Id: I5bb40f2083cf4509f7344fd06f8d9581283a7e74
Signed-off-by: Keith Smith <bksmith@us.ibm.com>