[ FAB-3581 ] Updated error message for multica fvt
The error message that changed as part of Change 11571,
FAB-3581 - Improve error handling (#3)
pdated a message that affected the outcome of the multica
test. The message in the test should be changed.

The test should have been changed as part of the
update, but for some reason the test did not run in CI because
it did non exiest in the tree by the time of the last patch
for 11571.

Multica test was merged in commit
084c6dc4e3613f9d152ecea09b2780f6fb4add69

and does in fact appear in the merge of FAB-3581:
baa0df401510aea63a7acf670b824f12345f484f

but does not appear in the tree for the last commit o
11571:
9895f6b15400819811b5c378194aca380fc31503

Change-Id: I6696cc9e1d4ca0ce1b91bccc2936dc324951ed04
Signed-off-by: Allen Bailey <eabailey@us.ibm.com>