FAB-5794 Mask credentials in debug messages
(backport for 1.0.2)

While this would not normally have been an
issue, there are several cases where 3rd
party operators are hosting/running
fabric-ca.  In those cases, the people
who monitor the service should not be able
to see any credentials in the logs if they
need to turn on debug logging to
troubleshoot.

Change-Id: I51aa67a80094278e53481a91b8a9252655bd603e
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>