[ FAB-6337 ] Update BASEREL version in Makefile
change baseimage from 0.3.1 to 0.4.2

For some reason, the TLS hostname check on the 0.4.2 is
enforcing localhost in lieu of IP address. This was causing
the TLS setup in the intermediateca_test to fail. Updated 
the test to include 'localhost' in the root CA TLScertificate,
as well as the IP address.

Change-Id: I8c6b7ad2d681e50d6a9071b98839d480d68c0cdd
Signed-off-by: Allen Bailey <eabailey@us.ibm.com>