[FAB-5300] Fix test cases failing on Mac
Some test cases that were added in the change set
https://gerrit.hyperledger.org/r/c/13705/ are failing
on Mac. They are failing because these test cases
depend on os.Getwd returning error after current
working directory is deleted. This is not happening
on Mac. On Mac, os.Getwd does not return an error as
the deleted current working directory is moved to
/private directory. This change set removes the
failing test cases.

Change-Id: Ia0485a810fc9272e5b4b99fe6d0e7a257bf6d50a
Signed-off-by: Anil Ambati <aambati@us.ibm.com>