[FAB-6561] Close DB when initCA fails
This CR makes it so that in case initCA returns an error
after opening the DB, newCA closes the DB before returning
a nil and an error and losing the reference to the DB forever.
Without the reference to the DB, it becomes impossible
to close the DB anymore. Trying to delete the DB file leads to
an error on vagrant/Windows:

ca_test.go:722: Remove failed: remove ../testdata/fabric-ca-server.
db: text file busy
FAIL

With this CR this error goes away.

Patch-set #2: rebase

Change-Id: Ic2cab315d0006a39e191358151d2ca44b6de5e83
Signed-off-by: Arnaud J Le Hors <lehors@us.ibm.com>