[FAB-6662] Make enroll attrs required by default
Currently, the AttributeRequest struct has "Require" attribute.
It is set to false if the user does not specify the attribute.
But the fabric-ca-client command line considers the enrollment
attributes as required if the attributes are not qualified
with ":opt".

This change set makes the API and command line client consistent.
The "Require" attribute of AttributeRequest struct has been changed
to "Optional", whose value is false by default. This makes the
specified enrollment attributes required by default.

Change-Id: I4710174a3e1d7a5a2026c7a482d77219d337e6a2
Signed-off-by: Anil Ambati <aambati@us.ibm.com>