[FAB-5782] Initialization failure on Postgres
PostgreSQL always expects to connect to a database, it does not
allow establishing a connection just to the server itself. This
causes issues when you specify a database in the connection string
that does not exist.

A default PostgresSQL installation comes with two default database
that can be used to establish a connection - "template1" and "postgres".
Connection retry logic is implemented  as defined below:

1. Connect to database specified in connection string
2. If 1 fails, connect to "postgres" database
3. If 2 fails, connect to "template1" database
4. If 1-3 fail, throw an error indicating database initialization failed
and that database needs to be created manually. The server will continue to
start but will try to reinitialize the database on the next request the server
receives.

Change-Id: I84f3d52a2b770ec1002fe8715f31cd1dfc9c2e5b
Signed-off-by: Saad Karim <skarim@us.ibm.com>