[FAB-6991] Fix max enrollments for bootstrap user
The max enrollments for the bootstrap user should use the
max enrollments setting of the server.  This is needed in order
to allow starting the server to enforce one-time passwords and
also use the default bootstrap user settings.

The additional test case failed prior to this change set,
but now succeeds.

Change-Id: I58264542689ad1589cf764696353e33924f063e5
Signed-off-by: Keith Smith <bksmith@us.ibm.com>
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>