[FAB-7291] Set default value for Chainfile attribute
In a fabric CA cluster that has two servers, if the cluster members are
started in parallel, there is a chance one server will initialize before
the other. When this happens, first server will enroll and get a CA cert,
private key and a CA chain and store them on the shared file system. The
second server will skip the enrollment since enrollment was done by the
first server, so getCACert method is not invoked. Since the
cfg.CA.Chainfile variable is set in the getCACert method, its value will
remain as an empty string. This causes an error when getcacerts request
is sent to the second server.

Fix is to set the cfg.CA.Chainfile to default value "ca-chain.pem" if a
value is not specified by the user. With this change, cluster fvt test
passes.

This problem was introduced by the change set:
https://gerrit.hyperledger.org/r/c/15373/. The problem was uncovered by
cluster_test.sh fvt script. This script did not exist when first patch set
for 15373 was submitted. Since there was no conflict between these two
changes, the Gerrit did not complain.

Change-Id: I1eb8db05fc9f79b7cbb989188c4f65bb8655ded1
Signed-off-by: Anil Ambati <aambati@us.ibm.com>