[FAB-8475] Run migration logic only if db level is old
Currently, DB migration logic is run during start up even if the
DB level is same as server level. With this change set, it will
only be run if the DB level is less than server level. Before this
change set logic was:
Update database schema, add users/affiliations from the config file and
migrate the data

After this change set:
Database schema is updated and data is migrated in one transaction and
then users/affiliations defined in the config file are added.

Change-Id: I473453969b034f039dec65ac26d950f1114750c5
Signed-off-by: Anil Ambati <aambati@us.ibm.com>
Signed-off-by: Saad Karim <skarim@us.ibm.com>