[FAB-10411] Use default version when not set
The metadata.Version variable specifies the Fabric CA version
and is set at build time, only if the server executable is built
using make command. But if the executable is built using
'go get -u github.com/hyperledger/fabric-ca/cmd/...' command the
Version variable is not set. When Version is not set, the server
panics and exits with an error.

This change sets a default value for Version variable, which
must be updated every time BASE_VERSION Makefile variable is updated.

Change-Id: I4104546bbc858c2e2473f1a23a141a89ad3c5194
Signed-off-by: Anil Ambati <aambati@us.ibm.com>