[FABC-882] Remove Fabric BaseImage
Replaces the use of the Fabric baseimage with a
raw golang container for building the fabric binaries
and debain:buster for the FVT test suite.

Fabric baseimage is a bloated, specialty image
that was historically used by Fabric with all
of dependent utilies require by fabric baked into it.
In the case of the CA FVT container, the only piece
we gain by using base image is having openssl already
installed in the image.

Signed-off-by: Brett Logan <Brett.T.Logan@ibm.com>