Fix broken URL links
Some links refer to the Fabric master branch, both in the GitHub
repository and in the published documentation. The GitHub branch is now
main, and the documentation is categorised by version, with latest
replacing master.

Added a replacement pattern in the doc configuration to specify the
FABRIC_VERSION of the Fabric documentation. References to the Fabric
documentation in the generated docs now use this replacement pattern.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>